Focal statistics refactoring part I

+++ pyspatialstats/grouped_stats/core/count.pxd
@@ -1,2 +1,2 @@
 cdef size_t _define_max_ind(size_t[:] ind) noexcept nogil
-cdef long* _grouped_count(size_t[:] ind, float[:] v, size_t max_ind) nogil
+cdef size_t* _grouped_count(size_t[:] ind, double[:] v, size_t max_ind) nogil
