+++ pyspatialstats/grouped_stats/core/max.pxd
@@ -1,1 +1,1 @@
-cdef float* _grouped_max(size_t[:] ind, float[:] v, size_t max_ind) nogil
+cdef double* _grouped_max(size_t[:] ind, double[:] v, size_t max_ind) nogil
