+++ pyspatialstats/grouped_stats/core/min.pxd
@@ -1,1 +1,1 @@
-cdef float* _grouped_min(size_t[:] ind, float[:] v, size_t max_ind) nogil
+cdef double* _grouped_min(size_t[:] ind, double[:] v, size_t max_ind) nogil
