+++ pyspatialstats/grouped_stats/core/std.pxd
@@ -1,1 +1,1 @@
-cdef float* _grouped_std(size_t[:] ind, float[:] v, float *mean_v, size_t max_ind) nogil
+cdef double* _grouped_std(size_t[:] ind, double[:] v, double *mean_v, size_t max_ind) nogil
